backend/requirements.txt
# Django
Django>=5.0,<5.1
djangorestframework>=3.14
django-cors-headers>=4.3
django-filter>=23.5
django-environ>=0.11

# Database
psycopg[binary]>=3.1

# Async Tasks
celery>=5.3
redis>=5.0

# SIFEN / XML
lxml>=5.1
xmlsec>=1.3
cryptography>=41.0
signxml>=3.2

# Utilities
python-dateutil>=2.8
pydantic>=2.5
httpx>=0.26

# Development
pytest>=7.4
pytest-django>=4.7
black>=23.12
ruff>=0.1
